+++ xchange-cryptotrade/src/main/java/com/xeiam/xchange/cryptotrade/dto/CryptoTradeTickersDeserializer.java
@@ -1,8 +1,9 @@
 package com.xeiam.xchange.cryptotrade.dto;
 
 import java.io.IOException;
-import java.math.BigDecimal;
+import java.util.HashMap;
 import java.util.Iterator;
+import java.util.Map;
 import java.util.Map.Entry;
 
 import com.fasterxml.jackson.core.JsonParser;
@@ -13,58 +14,29 @@ import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.xeiam.xchange.cryptotrade.dto.marketdata.CryptoTradeTicker;
+import com.xeiam.xchange.currency.CurrencyPair;
+import com.xeiam.xchange.utils.jackson.CurrencyPairDeserializer;
 
-public class CryptoTradeTickersDeserializer extends JsonDeserializer<CryptoTradeTicker> {
-
-  private static BigDecimal getNumberIfPresent(final String numberString) {
-
-    return numberString.isEmpty() ? null : new BigDecimal(numberString);
-  }
-
-  private static CryptoTradeTicker getTickerFromJsonNode(JsonNode tickerDataNode, JsonNode statusNode) {
-
-    final BigDecimal last = getNumberIfPresent(tickerDataNode.path("last").asText());
-    final BigDecimal low = getNumberIfPresent(tickerDataNode.path("low").asText());
-    final BigDecimal high = getNumberIfPresent(tickerDataNode.path("high").asText());
-    final BigDecimal minAsk = getNumberIfPresent(tickerDataNode.path("min_ask").asText());
-    final BigDecimal maxBid = getNumberIfPresent(tickerDataNode.path("max_bid").asText());
-    BigDecimal volumeTradeCurrency = null;
-    BigDecimal volumePriceCurrency = null;
-    if (tickerDataNode instanceof ObjectNode) {
-      final ObjectNode tickerDataObjectNode = (ObjectNode) tickerDataNode;
-      final Iterator<Entry<String, JsonNode>> tickerDataFields = tickerDataObjectNode.fields();
-      while (tickerDataFields.hasNext()) {
-        final Entry<String, JsonNode> tickerDataEntry = tickerDataFields.next();
-        if (tickerDataEntry.getKey().startsWith("vol_")) {
-          if (volumeTradeCurrency == null)
-            volumeTradeCurrency = getNumberIfPresent(tickerDataEntry.getValue().asText());
-          else {
-            volumePriceCurrency = getNumberIfPresent(tickerDataEntry.getValue().asText());
-            break;
-          }
-        }
-      }
-    }
-
-    final String status = statusNode != null ? statusNode.path("status").asText() : null;
-    final String error = statusNode != null ? statusNode.path("error").asText() : null;
-
-    return new CryptoTradeTicker(last, low, high, volumeTradeCurrency, volumePriceCurrency, minAsk, maxBid, status, error);
-  }
-
-  public static CryptoTradeTicker getTickerFromJsonNode(JsonNode tickerDataNode) {
-
-    return getTickerFromJsonNode(tickerDataNode, null);
-  }
+public class CryptoTradeTickersDeserializer extends JsonDeserializer<Map<CurrencyPair, CryptoTradeTicker>> {
 
   @Override
-  public CryptoTradeTicker deserialize(JsonParser jp, DeserializationContext ctxt) throws IOException, JsonProcessingException {
+  public Map<CurrencyPair, CryptoTradeTicker> deserialize(JsonParser jp, DeserializationContext ctxt) throws IOException, JsonProcessingException {
 
     final ObjectCodec oc = jp.getCodec();
-    final JsonNode statusNode = oc.readTree(jp);
-    final JsonNode tickerDataParentNode = statusNode.path("data");
+    final JsonNode rootNode = oc.readTree(jp);
 
-    return getTickerFromJsonNode(tickerDataParentNode, statusNode);
+    final Map<CurrencyPair, CryptoTradeTicker> tickers = new HashMap<CurrencyPair, CryptoTradeTicker>();
+    if (rootNode.size() == 1) {
+      final ObjectNode dataNode = (ObjectNode) rootNode.get(0);
+      final Iterator<Entry<String, JsonNode>> tickersArray = dataNode.fields();
+      while (tickersArray.hasNext()) {
+        final Entry<String, JsonNode> tickerData = tickersArray.next();
+        final CurrencyPair pair = CurrencyPairDeserializer.getCurrencyPairFromString(tickerData.getKey());
+        final CryptoTradeTicker ticker = CryptoTradeTickerDeserializer.getTickerFromJsonNode(tickerData.getValue());
+        tickers.put(pair, ticker);
+      }
+    }
+    return tickers;
   }
 
 }
